Add some scripts for creating videos of the programs.
Each of the programs here can be run through RISC OS and recorded
to make a video of its operation. These videos are generally quite
small, and I've limited them to just 2 minutes. Not all of the
programs are able to be recorded in this way - some need key input,
some don't work (probably Pyromaniac bugs), and some need pointer
input.